+++ X10D/src/Math/ByteExtensions.cs
@@ -1,6 +1,4 @@
 using System.Diagnostics.Contracts;
-using System.Runtime.CompilerServices;
-using X10D.CompilerServices;
 
 namespace X10D.Math;
 
@@ -20,20 +18,5 @@ public static class ByteExtensions
     public static bool IsPrime(this byte value)
     {
         return ((long)value).IsPrime();
-    }
-
-    /// <summary>
-    ///     Returns the multiplicative persistence of a specified value.
-    /// </summary>
-    /// <param name="value">The value whose multiplicative persistence to calculate.</param>
-    /// <returns>The multiplicative persistence.</returns>
-    /// <remarks>
-    ///     Multiplicative persistence is defined as the recursive digital product until that product is a single digit.
-    /// </remarks>
-    [Pure]
-    [MethodImpl(CompilerResources.MaxOptimization)]
-    public static int MultiplicativePersistence(this byte value)
-    {
-        return ((long)value).MultiplicativePersistence();
     }
 }