+++ X10D/src/Math/UInt16Extensions.cs
@@ -23,19 +23,4 @@ public static class UInt16Extensions
     {
         return ((ulong)value).IsPrime();
     }
-
-    /// <summary>
-    ///     Returns the multiplicative persistence of a specified value.
-    /// </summary>
-    /// <param name="value">The value whose multiplicative persistence to calculate.</param>
-    /// <returns>The multiplicative persistence.</returns>
-    /// <remarks>
-    ///     Multiplicative persistence is defined as the recursive digital product until that product is a single digit.
-    /// </remarks>
-    [Pure]
-    [MethodImpl(CompilerResources.MaxOptimization)]
-    public static int MultiplicativePersistence(this ushort value)
-    {
-        return ((ulong)value).MultiplicativePersistence();
-    }
 }